LayoutTests/fast/profiler/anonymous-function-calls-eval-expected.txt

CONSOLE MESSAGE: line 59: Anonymous function calles eval
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  (program) anonymous-function-calls-eval.html 8
CONSOLE MESSAGE: line 75:   (program) anonymous-function-calls-eval.html 4
CONSOLE MESSAGE: line 75:   onload anonymous-function-calls-eval.html 21
CONSOLE MESSAGE: line 75:    startTest anonymous-function-calls-eval.html 10
CONSOLE MESSAGE: line 75:     (anonymous function) anonymous-function-calls-eval.html 11
CONSOLE MESSAGE: line 75:      eval  0
CONSOLE MESSAGE: line 75:       (program)  1
CONSOLE MESSAGE: line 75:        insertNewText profiler-test-JS-resources.js 17
CONSOLE MESSAGE: line 75:         createElement  0
CONSOLE MESSAGE: line 75:         createTextNode  0
CONSOLE MESSAGE: line 75:         appendChild  0
CONSOLE MESSAGE: line 75:         getElementById  0
CONSOLE MESSAGE: line 75:     endTest profiler-test-JS-resources.js 1
CONSOLE MESSAGE: line 75:  (idle)  0
This page's JavaScript calls an anonymous function which calls eval(). 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. Beneath onload and startTest it should show an (anonymous function) which has a (program) child.
This is inserted Text
